backend/requirements.txt
# Production: Azure App Service (Python 3.11), PostgreSQL, Gunicorn
# Install: pip install -r requirements.txt
setuptools>=65.0.0
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-allauth==0.57.0
django-otp==1.2.4
django-cryptography==1.1
cryptography==41.0.7
pycryptodome==3.19.0
psycopg[binary]>=3.1.0
django-environ==0.11.2
djangorestframework-simplejwt==5.3.0
requests>=2.28.0
PyJWT[crypto]>=2.6.0
drf-spectacular==0.26.5
django-auditlog==3.0.0
python-json-logger==2.0.7
django-ratelimit==4.1.0
# Access DB import (management command import_from_access)
pyodbc>=5.0.0
python-dateutil==2.8.2
pytz==2023.3
gunicorn==21.2.0
whitenoise==6.6.0
azure-keyvault-keys==4.8.0
azure-identity==1.15.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
attrs>=22.0.0